Added button to create new screens
Created a button at the top of the page to create a user selected size
of new pixels to draw on, incomplete as the width of the screen is
fixed at 4 squares wide and only extends in the vertical axis. pixels
should be scaling in size to take up the width.